chore: refactor component imports
