+++ src/carbon-components-svelte/CodeSnippet/CodeSnippet.svelte
@@ -95,9 +95,9 @@
   export let ref = null;
 
   import { createEventDispatcher, tick, onMount } from "svelte";
-  import ChevronDown from "../../icons/ChevronDown.svelte";
+  import ChevronDown from "../icons/ChevronDown.svelte";
   import Button from "../Button/Button.svelte";
-  import CopyButton from "../../CopyButton/CopyButton.svelte";
+  import CopyButton from "../CopyButton/CopyButton.svelte";
   import CodeSnippetSkeleton from "./CodeSnippetSkeleton.svelte";
 
   const dispatch = createEventDispatcher();
