+++ src/carbon-components-svelte/DataTable/DataTable.svelte
@@ -129,7 +129,7 @@
   import { createEventDispatcher, setContext } from "svelte";
   import { writable } from "svelte/store";
   import ChevronRight from "../icons/ChevronRight.svelte";
-  import InlineCheckbox from "../carbon-components-svelte/Checkbox/InlineCheckbox.svelte";
+  import InlineCheckbox from "../Checkbox/InlineCheckbox.svelte";
   import RadioButton from "../RadioButton/RadioButton.svelte";
   import Table from "./Table.svelte";
   import TableBody from "./TableBody.svelte";
